Upload images into 2 separate folders in Firebase: - "dev" if environment variable REACT_APP_DEV is evaluated to true. - "public" otherwise.
