+++ todo.txt
@@ -1,0 +1,3 @@
+How to easily retrieve data (image AND associated metadata) from Firebase ?
+
+How to separe uploaded test/dev images and production/public images into Firebase ?
